+++ packages/tsurlfilter/src/utils/utils.ts
@@ -1,3 +1,5 @@
+// TODO rename filename from utils.ts to string-utils.ts
+
 /**
  * Splits the string by the delimiter, ignoring escaped delimiters.
  *